Rename LocalVideoUploadProgress to VideosToUpload and update DbSet reference in VideosDbContext
