+++ src/TB.DanceDance.Mobile/Data/VideosDbContext.cs
@@ -1,7 +1,4 @@
 ﻿using Microsoft.EntityFrameworkCore;
-using System.ComponentModel.DataAnnotations;
-using TB.DanceDance.API.Contracts.Responses;
-using TB.DanceDance.Mobile.Data.VideoModels;
 
 namespace TB.DanceDance.Mobile.Data;
 
@@ -12,5 +9,5 @@ public class VideosDbContext : DbContext
 
     }
 
-    public DbSet<LocalVideoUploadProgress> LocalVideoUploadProgresses { get; set; }
+    public DbSet<VideosToUpload> LocalVideoUploadProgresses { get; set; }
 }
